Check overlap between evaled and toEval
Previously, this code did not consider the fact that a number to
be evaluated would, in the next iteration, be inserted into evaled,
thus changing the indexing for any element in evaled after that
element.

Accordingly, once the algorithm reached >=n elements, it would return
the nth element, even though that element would probably change in
subsequent iterations.

Thus, check what the front of the toEval deque is, since that will
determine how many elements within evaled are invariant at that index.