Implement clever pq_only solution
from https://www.codewars.com/kata/reviews/578cf68aa684bd8ed3000885/groups/57cdea6efa9fc5d6140001d9

The idea behind this solution is that we don't actually care about our
invariants; once we evaluate 1 (to obtain 3 and 4), we can toss 1
aside, so long as we know know that there is a single invariant.
Then, once we evaluate 3 (to obtain 7 and 10) and 4 (to obtain
9 and 13), we can toss 3 and 4 aside, literally not caring what they were,
so long as we now know that we have three total invariants.

This matters because we care about the nth invariant.  Once we have
n-1 invariants, we know that the next one in a sorted list of
numbers to evaluate will be our nth invariant.

Sorting here is key though, and is provided by the C++
priority_queue data structure, which allows FIFO queue behavior,
but which guarantees that the elements are sorted (here,
in ascending order).

Accordingly, although the numbers we are pushing to the pq
are blended in with the numbers which are already there, the pq
ensures that they are sorted, thus that the top() is the next
invariant to evaluate, and thus we can just do n-1 iterations
of popping and pushing two new numbers, secure in the knowledge
that the next number to pop will be the nth invariant
(That is, the nth term in the Twice Linear sequence this problem
establishes).